# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Musholla Nurul Iman API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface Page {
  id: number;
  title: string;
  slug: string;
  content: string;
  isPublished: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePage {
  title: string;
  slug: string;
  content: string;
  isPublished: boolean;
}

export interface Menu {
  id: number;
  label: string;
  url: string;
  order: number;
  parentId?: number | null;
  isActive: boolean;
}

export interface CreateMenu {
  label: string;
  url: string;
  order: number;
  parentId?: number | null;
  isActive: boolean;
}

export interface News {
  id: number;
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  imageUrl?: string | null;
  author: string;
  isPublished: boolean;
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNews {
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  imageUrl?: string | null;
  author: string;
  isPublished: boolean;
  publishedAt?: string | null;
}

export interface Class {
  id: number;
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  imageUrl?: string | null;
  author: string;
  isPublished: boolean;
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClass {
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  imageUrl?: string | null;
  author: string;
  isPublished: boolean;
  publishedAt?: string | null;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  location: string;
  startDate: string;
  endDate?: string | null;
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateEvent {
  title: string;
  description: string;
  location: string;
  startDate: string;
  endDate?: string | null;
  imageUrl?: string | null;
  isActive: boolean;
}

export interface GalleryItem {
  id: number;
  type: string;
  title: string;
  imageUrl: string;
  category: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateGalleryItem {
  type: string;
  title: string;
  imageUrl: string;
  category: string;
  isActive: boolean;
}

export interface CctvCamera {
  id: number;
  name: string;
  location: string;
  streamUrl: string;
  embedUrl?: string | null;
  isActive: boolean;
  order: number;
  description?: string | null;
}

export interface CreateCctvCamera {
  name: string;
  location: string;
  streamUrl: string;
  embedUrl?: string | null;
  isActive: boolean;
  order: number;
  description?: string | null;
}

export interface SiteSettings {
  siteName: string;
  tagline: string;
  description: string;
  address: string;
  phone: string;
  email: string;
  facebook?: string | null;
  twitter?: string | null;
  instagram?: string | null;
  youtube?: string | null;
  logoUrl?: string | null;
  heroImageUrl?: string | null;
  heroTitle: string;
  heroSubtitle: string;
}

export interface PrayerTimes {
  fajr: string;
  dhuhr: string;
  asr: string;
  maghrib: string;
  isha: string;
  jumuah: string;
  updatedAt?: string | null;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token?: string | null;
  message: string;
}

export type GetNewsParams = {
  published?: boolean;
};

export type GetClassesParams = {
  published?: boolean;
};