# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Musholla Nurul Iman API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: getClasses
      tags: [classes]
      summary: Get all classes
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClass"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      operationId: getClassById
      tags: [classes]
      summary: Get class by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    put:
      operationId: updateClass
      tags: [classes]
      summary: Update class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClass"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /cctv:
    get:
      operationId: getCctvCameras
      tags: [cctv]
      summary: Get all CCTV cameras
      responses:
        "200":
          description: List of cameras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CctvCamera"
    post:
      operationId: createCctvCamera
      tags: [cctv]
      summary: Create CCTV camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCctvCamera"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CctvCamera"
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
  /prayer-times:
    get:
      operationId: getPrayerTimes
      tags: [prayer]
      summary: Get prayer times
      responses:
        "200":
          description: Prayer times
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerTimes"
    put:
      operationId: updatePrayerTimes
      tags: [prayer]
      summary: Update prayer times
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrayerTimes"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerTimes"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Class:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
        imageUrl:
          type: string
          nullable: true
        author:
          type: string
        isPublished:
          type: boolean
        publishedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - content
        - excerpt
        - author
        - isPublished
        - createdAt
        - updatedAt
    CreateClass:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
        imageUrl:
          type: string
          nullable: true
        author:
          type: string
        isPublished:
          type: boolean
        publishedAt:
          type: string
          nullable: true
      required:
        - title
        - slug
        - content
        - excerpt
        - author
        - isPublished
...
    CctvCamera:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        streamUrl:
          type: string
        embedUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        order:
          type: integer
        description:
          type: string
          nullable: true
      required:
        - id
        - name
        - location
        - streamUrl
        - isActive
        - order
    CreateCctvCamera:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        streamUrl:
          type: string
        embedUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        order:
          type: integer
        description:
          type: string
          nullable: true
      required:
        - name
        - location
        - streamUrl
        - isActive
        - order
    SiteSettings:
      type: object
      properties:
        siteName:
          type: string
        tagline:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        facebook:
          type: string
          nullable: true
        twitter:
          type: string
          nullable: true
        instagram:
          type: string
          nullable: true
        youtube:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          nullable: true
        heroTitle:
          type: string
        heroSubtitle:
          type: string
      required:
        - siteName
        - tagline
        - description
        - address
        - phone
        - email
        - heroTitle
        - heroSubtitle
    PrayerTimes:
      type: object
      properties:
        fajr:
          type: string
        dhuhr:
          type: string
        asr:
          type: string
        maghrib:
          type: string
        isha:
          type: string
        jumuah:
          type: string
        updatedAt:
          type: string
          nullable: true
      required:
        - fajr
        - dhuhr
        - asr
        - maghrib
        - isha
        - jumuah
...