# lib/api-zod/src/generated/types/createClass.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Musholla Nurul Iman API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClass {
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  imageUrl?: string | null;
  author: string;
  isPublished: boolean;
  publishedAt?: string | null;
}